+++ addons/mail/models/__init__.py
@@ -47,6 +47,7 @@ from . import ir_http
 from . import ir_model
 from . import ir_model_fields
 from . import ir_ui_view
+from . import ir_qweb
 from . import res_company
 from . import res_config_settings
 from . import res_partner
