+++ odoo/addons/test_assetsbundle/tests/test_assetsbundle.py
@@ -21,7 +21,7 @@ from odoo.addons.base.models.ir_attachment import IrAttachment
 from odoo.modules.module import get_resource_path, get_manifest
 from odoo.tests import HttpCase, tagged
 from odoo.tests.common import TransactionCase
-from odoo.addons.base.models.qweb import QWebException
+from odoo.addons.base.models.ir_qweb import QWebException
 from odoo.tools import mute_logger, func
